pyproject.toml
[project]
name = "daily-feed"
version = "0.1.0"
description = "Daily feed summarizer CLI"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "Daily Feed" }]

dependencies = [
  "typer>=0.12.0",
  "rich>=13.0.0",
  "httpx>=0.26.0",
  "pyyaml>=6.0.0",
  "trafilatura>=1.6.0",
  "readability-lxml>=0.8.1",
  "beautifulsoup4>=4.12.0",
  "rapidfuzz>=3.6.0",
  "jinja2>=3.1.0",
  "python-dotenv>=1.0.0",
]

[project.scripts]
daily-feed = "daily_feed.cli:app"

[build-system]
requires = ["setuptools>=68.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["daily_feed", "daily_feed.providers"]
include-package-data = true

[tool.setuptools.package-data]
daily_feed = ["templates/*.html"]
